Updates to Urban 1D including test_all_cllasses
