+++ __temp.gxy
@@ -1,0 +1,71 @@
+[QTBDY__a]
+X=-113985.064
+Y=219232.026
+
+[BLOCKAGE__a]
+X=-90259.574
+Y=215637.255
+
+[CONDUIT_CIRCULAR_b]
+X=-80014.476
+Y=210964.052
+
+[CONDUIT_CIRCULAR_ds]
+X=-67291.752
+Y=207587.179
+
+[NCDBDY__ds]
+X=-52419.906
+Y=207928.406
+
+[Connections]
+ConnectionCount=3
+1=BLOCKAGE__a,CONDUIT_CIRCULAR_b
+2=CONDUIT_CIRCULAR_b,CONDUIT_CIRCULAR_ds
+3=CONDUIT_CIRCULAR_ds,NCDBDY__ds
+
+[VISTEXT]
+VISTEXTCOUNT=0
+
+[Font]
+Font name=Tahoma
+Charset=1
+Height=-11
+Bold=0
+Italic=0
+Underline=0
+Strikeout=0
+
+[Text]
+Layout=1
+
+[Display]
+MiniNodes=0
+MiniStructureNodes=0
+NodeLabels=0
+Connections=1
+
+[General]
+UnitMove=0
+ExportConnections=0
+GeneralConnectionsColour=0
+LateralConnectionsColour=16711680
+SpillConnectionsColour=8421504
+
+[Grid]
+Show=0
+X=25
+Y=25
+
+[RWC]
+Top=273333.333333333
+Left=-166468.723954906
+Right=90557.4198359433
+Bottom=126666.666666667
+
+[PC]
+Top=0
+Left=0
+Right=0
+Bottom=0
+
