+++ docs/source/user_guide/index.rst
@@ -12,6 +12,7 @@ can be used, code reference and examples.
    zzn
    ied
    dat
+   inp
    units
