+++ floodmodeller_api/units/boundaries.py
@@ -20,7 +20,7 @@ from ._base import Unit
 from .helpers import (join_10_char, join_12_char_ljust, join_n_char_ljust,
                       split_10_char, split_12_char, split_n_char, _to_str, _to_float, _to_int,
                       _to_data_list)
-from .validation import _validate_unit, parameter_options
+from floodmodeller_api.validation import _validate_unit
