+++ floodmodeller_api/units/sections.py
@@ -19,7 +19,7 @@ import pandas as pd
 from ._base import Unit
 from .helpers import (join_10_char, join_12_char_ljust, join_n_char_ljust,
                       split_10_char, split_12_char, split_n_char, _to_float, _to_int)
-from .validation import _validate_unit, parameter_options
+from floodmodeller_api.validation import _validate_unit
 
 
 class RIVER(Unit):
