+++ floodmodeller_api/units/structures.py
@@ -20,7 +20,7 @@ import pandas as pd
 from ._base import Unit
 from .helpers import (join_10_char, join_12_char_ljust, join_n_char_ljust,
                       split_10_char, split_12_char, split_n_char, _to_float, _to_str, _to_int, _to_data_list)
-from .validation import _validate_unit, parameter_options
+from floodmodeller_api.validation import _validate_unit
 
 class CULVERT(Unit):
     """ The CULVERT class supports two culvert sub-types in Flood Modeller: INLET and OUTLET. Each of these sub-types forms
