+++ setup.py
@@ -8,7 +8,12 @@ setup(
    version='0.3.12.post1',
    author='Jacobs',
    author_email='joe.pierce@jacobs.com',
-   packages=['floodmodeller_api', 'floodmodeller_api.units'],
+   packages=[
+       'floodmodeller_api',
+       'floodmodeller_api.units',
+       'floodmodeller_api.validation',
+       'floodmodeller_api.urban1d'
+       ],
    scripts=[],
    project_urls={
        'API Documentation': 'https://help.floodmodeller.com/api/',
